fix: fix data type error

+++ iOS/FIT5140Assignment3iOS/FIT5140Assignment3iOS/Network/RequestModel/SpeedRecord.swift
@@ -9,8 +9,8 @@ import UIKit
 
 class SpeedRecord: Codable {
     var id:String?
-    let currentSpeed:Int
-    let limitedSpeed:Int
+    let currentSpeed:Double
+    let limitedSpeed:Double
     let latitude:Double
     let longitude:Double
     let recordId:String?
